Bug 1198978 - Make sure the Save button is completely disabled when it is disabled.
ng-disabled only completely disables actual input elements, but we use it
on a fake button (actually a span), so the ng-click is still being called
even when the "button" is "disabled". This change fixes it so the ng-click
doesn't call the save() function unless the ng-disabled condition is NOT
met.