Bug 1160111 - Add ui for self-serve application management.
An authenticated user can access the new ui and add new applications. Only the owner of the application can see it in that ui.
When an application is created it gets automatically a secret key, which is available in the application details.
The application needs to be approved by a member of staff before it can be used for data submission; that needs to be done in the admin panel.
In the admin panel it is also possible to reset the secret key of one or more applications. The new secret will be available to its owner in the application details.