Bug 1205871 - Add support to transition completed jobs to retry
In buildbot, a job transitions directly from ``running`` to ``retry`` in the
cases where it will be retried automatically.  But for upcoming Buildbot
Bridge and Task Cluster jobs that we will ingest, we need to be able to make
the transition from ``completed`` to ``retry``.

This change involves special case handling that patches an already ugly
ingestion scenario.  When we no longer ingest from Buildbot directly, we
can refactor ``jobs.py`` to remove these work-arounds.  A bit part of that
is that there is no unique job id for a Buildbot job to know when it has
transitioned.  Each time a job is retried, it has the same request_id and
request_time.