Bug 1205412 - Fix highlighting of revisions in perf graphs
When we had two series with the same signature on different branches, we
would incorrectly highlight two points on the same graph instead of
highlighting the relevant revisions on each. This used to work because
we associated a branch with each series signature, so could identify
them uniquely before, but we don't do this any more. Fortunately the fix here
is simple: just use the previously defined series instead of trying to
look it up all over again.