Bug 1206108 - Separate `grunt build` npm packages from those for testing
To run grunt build as part of deploy, we're going to need to npm install
the dependencies required for it. However we do not need to install the
testing related packages, so by splitting the two between `dependencies`
and `devDependencies` we'll have more control. eg:
`npm install --production` will only install those under `dependencies`.

See:
https://docs.npmjs.com/files/package.json#dependencies
https://docs.npmjs.com/files/package.json#devdependencies
https://docs.npmjs.com/misc/config#production