Bug 1206108 - Add grunt option to skip loading `devDependencies` tasks
In production we plan to only install the packages listed under
`dependencies` in package.json. However by default load-grunt-tasks
loads tasks from packages from `devDependencies` too, which results in
warnings if those packages are not installed locally.

To ensure that the local development grunt tasks are still loaded, but
that warnings are not shown for missing packages on production, a new
option is added to select which packages are loaded. eg:
`grunt build --production`

See:
http://gruntjs.com/api/grunt.option
https://github.com/sindresorhus/load-grunt-tasks/blob/master/readme.md#only-load-from-devdependencies