Bug 1207237 - Fix cursor when hovering over non-selectable job group
We were applying the "pointer" style, which implied you could select
something when you couldn't before. We should only use that style when
hovering over the job buttons themselves. When hovering other parts
of the job group, use the "default" cursor which communicates that there
is nothing to click on or select.