Bug 1169320 - Adds ability to ingest Jobs from Pulse exchanges
This ingests jobs over a pulse stream if the resultset for the job
already exists.  There is a new management command called
``ingest_from_pulse`` which reads from a set of exchanges specified
in ``base.py`` or environment variables.  These will be loaded in a
local celery queue for ingestion into the DB.

The user is expected to create their own Pulse user with Pulse Guardian
to be able to ingest jobs as well as post to their own queue.

A JSON Schema in the form of a YML file is included to validate the jobs
prior to ingestion.  The resultset/push for a job must
already exist in Treeherder or the job will be skipped.