Bug 1211524 - Add readme badges for requires.io and david-dm.org
So we can more easily track out of date dependencies.

Note that since we pin our nodejs packages, the Node badges will remain
green even if dependencies are out of date, due to:
https://github.com/alanshaw/david-www/issues/48
...however the badge is still useful since it's just one click to see
the table.

To get the requires.io badge working a GitHub hook has to be set up,
which I've already done.