Bug 1212951 - Set 'verbose-name-plural' for models that need it
By default Django appends an 's' onto the singular model name (in
addition to munging CamelCase) when generating the plural form that is
used in the admin panel/elsewhere. For example:

Device -> "devices"
RepositoryGroup -> "repository groups"

For some of our model names this doesn't make sense and results in
items in the admin panel like "Credentialss", "Bugscaches" or
"Repositorys", so we use verbose-name-plural to explicitly set the name
that should be used:
https://docs.djangoproject.com/en/1.8/ref/models/options/#verbose-name-plural

"Credentials" is valid for the singular (ie: model name), according to:
http://english.stackexchange.com/questions/252073/which-is-correct-credential-or-credentials

"ReferenceDataSignatures" should really be renamed to the singular, but
for now I've just left a TODO and set verbose-name-plural so it looks
less awkward in the admin panel.