Bug 1079796 - store long and short revisions
We store both long and short, but only utilize the short (as before).  We
need to populate all the short and long revision records before we can
start using them.  So after this commit, we will begin backfilling the
old records that don't yet have those values populated.  Once they all
are, we can move to using the long_revision primarily in Bug 1199364.