Bug 1194830: Add a button to trigger new buildbot jobs
This adds a "Add new jobs" button that shows all buildbot jobs
that can be added as normal treeherder jobs (with a new CSS class).
After logged in users click on the jobs they want to trigger, they
can click on "Trigger new jobs" to send a pulse message that will
trigger the jobs.