Bug 1217374 - Log django.request WARNINGs on stage/prod
HTTP errors with a status code of 5XX are logged as ERRORs, and so were
already being output to the log on stage/prod. Status codes of 4XX are
logged as WARNINGs, and so are now logged too. This should make 403s
and 401s clearer when trying to debug issues with people submitting to
our API - particularly given they don't show up in New Relic either, due
to bug 1201086.