Bug 1185901 - Convert the bugscache table from MyISAM to InnoDB
MySQL 5.6 supports full text searching, so we no longer need to use
MyISAM. Django doesn't let us mix and match engines, so we need to pick
one and stick with it, for bug 1193836.

The bugscache table has already been updated manually on stage/prod and
Heroku; this switches it for Travis/the local Vagrant environment.