Bug 1221064 - improve job list performance
The job list endpoint was joining the job table and the resultset table
with the only purpose of sorting the results by push_timestamp. Sorting
by result_set_id seems to be a good enough solution and at the same time
it avoids running a poor performing join between job and resultset.