Bug 1221079 - make unclassified failures mode show "auto classified"
This also removes the ``classifiedState`` filter from the filter panel
and the ``jobFilters`` service because it is now dead code due to the
previous commit.  This was doing special magic around
whether a job was “classified” or not.  But it’s supplanted by just
using the field filter for ``failure_classification_id`` like we do
now.