Bug 1222025 - persist selected job in URL so it re-selects on page reload
If the url is used and the job is not found in the loaded result sets, then throw up an
error notification with a link to load the right resultset and selectedJob.  This could
happen if a job was selected, but no revision range provided and that job's resultset
was no longer in the "last 10" that get loaded by default.