Bug 1225154 - Remove out of date fig/docker configs
We will likely switch to a Docker based solution in the future, however
for now the configs are unused and out of date (fig has been replaced
with docker-compose, we no longer generate per-repo credentials etc).
In addition, there are tools that can convert the Heroku Procfile into
Docker configs, so we may end up using a solution based on that instead.

As such, let's remove the configs (they'll always be available in
version control history) to reduce the maintenance burden and clutter in
the repo.