Bug 1226565 - Stop update-parse-status from doing an additional SELECT
Previously it was performing both an UPDATE and then a SELECT, where the
SELECT was being performed on the read_only DB host. This meant in cases
of replication lag, the update-parse-status would return a 404, since
`get_job_log_url_detail` raises `ObjectNotFoundException` if zero rows
rows are found on the read_only host.

The SELECT was performing two tasks:
1) Allow the update-parse-status API to respond with the job_log_url
   record. However this is of little use to consumers of that endpoint.
2) Ensuring the API would return 404 if the job_log_url id in question
   actually did not exist at all (even on the master).

We can now perform (2) using datasource's new `return_type` of
`rowcount`, which means we only have to perform the UPDATE and not also
the SELECT.

This not only avoids the race when we're suffering from replication lag,
but also halves the number of queries run for each log parsed.