Bug 1229022 - Make treeherder-client HTTPError output less verbose
The request body, request headers & response headers are now output with
a log level of `debug` rather than `error`, to reduce verbosity of logs.

In addition, the response body has now been combined with the log line
that outputs the URL, to make it easier to correlate lines in the log.
The type in the `_get_json()` ("Error *submitting* data") has also been
fixed, and string substitution now occurs lazily.