Bug 1162947 - Move buildapi project validation to should_skip_project()
For each of the pending/running and complete codepaths we loop through
the dataset twice (once to generate the list of revisions to look up and
once to actually process the jobs). Previously we were missing the
project validation on some of these instances, but rather than duplicate
the code four times, let's move it to a new method.