Bug 1117583 - Add an API endpoint that creates a new bug in Bugzilla.
This endpoint receives a http POST request from the UI containing information about the bug to be filed (product, component, summary, version, description), then formats it properly as a submission to Bugzilla's REST API, using a server-side Bugzilla API key, and adding a "treeherder" comment tag and an "intermittent-failure" keyword.

The API then passes back either the bug ID (if the submission was successful) or Bugzilla's failure response if something went wrong.