Bug 1237096 - Fix to bad filtering of runnable jobs due to missing fields
Clicking the "Add new jobs" action menu item tries to add the runnable jobs,
but they are filtered out because they don't have some fields a normal job
has (like Tier, etc).  When there is a filter on a field, but a job object
doesn't have that field, it should be considered a 'pass' and show that job.