Bug 1157149,1240210 - The unclassified failure count should respect selected tiers of jobs
Keep track of the count of unclassified jobs currently loaded that match the
selected Tier settings.

When an event takes place that could change that value (new jobs loaded, classified
or unclassified) then update that count again.  This uses the unclassified map like
we did before, but re-evaluates it as needed.

This also fixes a bug where the interaction of the Tier dropdown and the Filters panel
didn't interact correctly.

This same technique can be leveraged (in the future) to get a count of unclassified
for the current filter set as well.