Bug 1233653 - Clarify the 'duration' for jobs that haven't started
For completed and running jobs, "duration" refers to the amount of time
the job has spent in the running state, and so intentionally does not
include the time spent queued (aka "pending").

However for jobs that haven't started, it's useful to display how long
they've been waiting (and this is what we already do) - we just need to
make it clear that it's not a 'duration' as normally referenced, but
instead the number of minutes the job has been queued.

Before:
> Duration: 10 minutes(s)

After:
> Duration: Not started (queued for 10 minute(s))