Bug 1239841 - clicking on the top level 'alerts' link yields no action
This happen when we already are viewing a specific alert id.

The problem is that ui-router merge the parameters from the current url
to determine the new parameters state. So the current is is merged, and
ui-router thinks that there is nothing to do since the states and
parameters are equivalent...

This fixes the issue, but it looks like a hack.