Bug 1210427 - Find the base line for a revision
In the perfherder compare view, selecting the new branch and revision
now search for the base line of that revision (e.g. the first branch and
revision available to compare to).

If the base line is found, the user can click a link to automatically
fill the base branch and revision.