Bug 1184795 - Stop hiding data ingestion exceptions.
With this change, if you get an exception during data ingestion in
DEBUG mode, it will simply throw the exception.  If not, it will
post the exception to New Relic with ``record_exception`` and move on
as we did before.  This will make it easier to detect errors in Vagrant
and prevent us from misssing them entirely on production and stage.