Add custom task synchronisation feature that uses the db backend and a counter.
celery doesn't implement this natively, which gives three choices:
1) Use ampq and lose messages at random
2) Start using redis and lose data on restart
3) Use the database backend and get an explosion of messages from polling

In practice I think the only reason the database backend has to poll is because using a simple counter
requires an atomic, threadsafe, get+set operation so that two tasks don't try to decrement the counter
at the same time, leaving it in an inconsistent state and so preventing the callback from running (e.g.
if A reads 2 and decrements to 1 in a non-atomic fashion, B might also read 2 and decrement to 1, leaving
the counter at 1 rather than 0).

For a MySQL (or Postgres) backend, this can be solved using SELECT ... FOR UPDATE, which locks the row
for the duration of the transaction so we can be sure that only one task tries to read the counter at a time.

Of course the implementation here is more constrained than the celery implementation, but we only have rather
simple needs.