Bug 1254673 - return resultset info when on API submission
This returns some basic info about the resultsets that were submitted via a POST to the API.
This is needed for Task Cluster as we transition away from ``revision_hash`` so that they
take it from us, instead of using what they submit.  This way, we can return them the
``revision_hash`` that we WANT to be moving to (PR 1322).  They also need this so they can
then send that new ``revision_hash`` to the tasks (jobs) that belong to it.