Bug 1242038 - Store unique job signatures for buildernames
If our regexes give the same platform values for several different
buildernames, then we may find the wrong signature and buildername for
those jobs.  This means that we will send the wrong buildername when
attempting to backfill or retrigger jobs.

This change ensures that the reference_data_signature is unique for
every buildername we get from buildbot.