Bug 1255406 - "Retrigger all" on a >10 pinned jobs only retriggers 10
The /artifacts endpoint defaults to only returning the first 10 items
unless a count is passed in.  We weren’t passing in that param, so we’d
only get back the first 10 (we need them for the buildbot request_ids).

This fixes it by passing in a “count” param equal to the number of jobs
we want to retrigger.