Bug 1257954 - Use parent signature property in import_perf_data
Modify the import_perf_data script to add each of a parent signatures'
subtest signatures, setting the parent_signature property appropriately.
Add a parent_signature argument to the _add_series function that
allows setting the parent_signature property on the
PerformanceSignature object to either None (if it is a parent)
or the hash of its parent (if it is a subtest signature).