Bug 1266116 - Add the Admin page: Full-page exclusion editor
This moves the Sheriff panel out to a full-page app.  The functionality
is basically the same, however.  Much of the logic and HTML were just
copied from the Sheriff panel.

This also introduces ReactJS to the repo.  This was used to speed up
the rendering of the reference data lists in the Exclusions detail
editor.  The ``reactselect`` component is defined in a .jsx file
and must be compiled with a new npm command called ``build-react``.