Bug 1264495 - honor visibility url param to see what jobs are excluded
Passes the ``exclusion_profile`` and ``visibility`` url params from
the main treeherder URL into the calls for jobs in each resultset.
With this, you can see exactly what will jobs are affected by the given
profile.

This also adds a button in the sheriff panel to quickly add these params
for each exclusion profile and open that in a new tab.

Also required a tiny fix for passing in an exclusion profile on a
repo that is not mentioned in that profile, but also passing in visibility
excluded.  It was just returning ALL jobs, where it should have
returned no jobs.