Bug 1266229 - Turn on ingestion of jobs from a taskcluster exchange
This contains several tweaks and fixes that allow us to ingest data from
a real Task Cluster owned exchange.

One of the main fixes is the way I was binding to the exchange and
routing keys with the same Queue.  Before, it was re-creating the Queue,
so would miss some of the bindings.