Bug 1267102 - Be more efficient about storing job title + search information
Since we only need it when actually filtering the job list, just dynamically
calculate there instead of keeping a (potentially large amount) of redundant
data in memory constantly. It doesn't take long to recalculate this relative
to other expenses (such as recalculating the DOM).

Since we were storing about 200 characters per job, and that a typical push
has upwards of 2000 jobs, this can save up to 2 megabytes on the default
treeherder view of 5 pushes.