Bug 1234247 - Only fetch missing pushlogs for completed jobs
The long-term aim is to make the fetch missing pushlogs tasks redundant,
by fixing the causes of missing pushes (if they're not already fixed).

However currently it's hard to identify whether it's still required,
since the standard pushlog task can race with job ingestion, causing us
to fetch 'missing' pushes that would have been otherwise retrieved fine.

By stopping the retrieval of missing pushes for pending/running jobs, we
give the standard pushlog retrieval time to run, prior to the job being
completed. If the push is still missing once we ingest the completed
job, we'll run the fetch missing pushes task as before - as a last
resort fall-back.