Bug 1100414 - Parse mozharness steps
Much of Firefox automation now lives in mozharness. Treeherder
currently parses buildbot and TaskCluster steps. This means
Treeherder's log parsing only shows a single step for mozharness.

The mozharness data is valuable. So let's parse it.

This commit introduces parsing of mozharness step start and
stop log entries.

One of the gotchas for mozharness is it only prints the time
in the logs - not the full date. So, we grab the date from
the last job, if available. This should "just work."