Bug 1273231 - Remove support for manually triggering a job parse
This was meant to reduce wait times for job parse status to be updated,
but it was accidentally broken over a year ago and no one noticed (we
started requiring hawk credentials, which were not available in the UI).
Let's just remove it. See bug 1076769 for details on the functionality
this was meant to provide.