Bug 1276355 - Fix IntegrityError for idx_revision_hash on resultset creation
When updating or inserting a resultset record, we need to check against
the revision_hash value because, while revision_hash is on its way out, we
do still use it as a fallback and there is a unique index for it.

This error can occur as a race condition when a job is being ingested that
does not yet have a resultset, but during ingestion, the resultset is
created in parallel.  So the process ingesting the job tries to insert
the resultset and gets the index violation.  This check will prevent that.