Bug 1163802 - For job cancellation only manually update pending jobs
If a buildbot job is cancelled whilst in the pending state, the
resultant "no longer running" job does not appear in builds-4hr (due to
buildbot limitations), which causes orphaned jobs in Treeherder.

To work around this, Treeherder pre-emptively updates the DB for jobs
when a user cancels them using the Treeherder interface. This is not
ideal, since the manual DB update can race with the job completing (if
the job finished before buildapi had time to cancel it), or worse can
show the job as cancelled even if the buildapi request failed to
complete, leaving the job as running even though Treeherder's UI will
show it as stopped.

We have to keep on performing this suboptimal workaround for pending
jobs (until we stop using buildbot), but we can at least limit it to
just the pending jobs case, since running jobs never needed this
workaround in the first place.